requirements.txt
python-telegram-bot==21.9
fastapi
uvicorn
requests
python-dotenv